readme.rdoc
Park Your Dog SF is an interactive dog park app for canine and human residents of San Francisco. With Park Your Dog SF, 
you'll be able to upload photos of your pup at play, write tips and reviews for your favorite or not so favortie dog parks,
and be kept up to date latest doggie meetups taking place in the city.